update: organized css stylesheets
